+++ src/stories/components/evo-sidebar.stories.ts
@@ -12,10 +12,15 @@ storiesOf('Components/Sidebar', module)
         ],
         providers: [
             EvoSidebarService,
-        ],
+        ]
     }))
     .add('default', () => ({
         component: EvoSidebarWrapperComponent,
+        moduleMetadata: {
+            declarations: [
+                EvoSidebarWrapperComponent,
+            ],
+        },
         props: {
             id: 'basket',
             header: 'Заголовок сайдбара',
@@ -25,6 +30,11 @@ storiesOf('Components/Sidebar', module)
     }))
     .add('with relativeFooter', () => ({
         component: EvoSidebarWrapperComponent,
+        moduleMetadata: {
+            declarations: [
+                EvoSidebarWrapperComponent,
+            ],
+        },
         props: {
             id: 'basket',
             header: 'Заголовок сайдбара',
@@ -35,6 +45,11 @@ storiesOf('Components/Sidebar', module)
     }))
     .add('with large size', () => ({
         component: EvoSidebarWrapperComponent,
+        moduleMetadata: {
+            declarations: [
+                EvoSidebarWrapperComponent,
+            ],
+        },
         props: {
             id: 'basket',
             size: 'large',