+++ src/stories/components/evo-toast.stories.ts
@@ -19,4 +19,9 @@ storiesOf('Components/Toast', module)
     }))
     .add('default', () => ({
         component: EvoToastWrapperComponent,
+        moduleMetadata: {
+            declarations: [
+                EvoToastWrapperComponent,
+            ],
+        }
     }));